chore: add plugin package with shims
