+++ pyproject.toml
@@ -10,6 +10,7 @@ package-mode = true
 packages = [
     { include = "entity", from = "src" },
     { include = "pipeline", from = "src" },
+    { include = "plugins", from = "src" },
     { include = "cli", from = "src" },
     { include = "config", from = "src" },
     { include = "registry", from = "src" },
