+++ src/pipeline/plugins/prompts/memory_retrieval.py
@@ -1,27 +1,3 @@
-from __future__ import annotations
+from plugins.prompts.memory_retrieval import MemoryRetrievalPrompt
 
-from typing import List
-
-from pipeline.context import ConversationEntry, PluginContext
-from pipeline.plugins import PromptPlugin
-from pipeline.resources.memory_resource import MemoryResource
-from pipeline.stages import PipelineStage
-
-
-class MemoryRetrievalPrompt(PromptPlugin):
-    """Fetch past conversation from memory and append it to the context."""
-
-    dependencies = ["memory", "llm"]
-    stages = [PipelineStage.THINK]
-
-    async def _execute_impl(self, context: PluginContext) -> None:
-        memory: MemoryResource = context.get_resource("memory")
-        history: List[ConversationEntry] = await memory.load_conversation(
-            context.pipeline_id
-        )
-        for entry in history:
-            context.add_conversation_entry(
-                content=entry.content,
-                role=entry.role,
-                metadata=entry.metadata,
-            )
+__all__ = ["MemoryRetrievalPrompt"]
