+++ src/pipeline/plugins/tools/__init__.py
@@ -1,6 +1,8 @@
-from .calculator_tool import CalculatorTool
-from .search_tool import SearchTool
-from .weather_api_tool import WeatherApiTool
+from plugins.tools import (
+    CalculatorTool,
+    SearchTool,
+    WeatherApiTool,
+)
 
 __all__ = [
     "CalculatorTool",
