+++ src/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Built-in plugin implementations."""
