+++ src/plugins/classifier.py
@@ -1,0 +1,3 @@
+from ..interfaces import PluginAutoClassifier
+
+__all__ = ["PluginAutoClassifier"]
