+++ src/plugins/prompts/__init__.py
@@ -1,0 +1,23 @@
+"""Expose built-in prompt plugins."""
+
+from .chain_of_thought import ChainOfThoughtPrompt
+from .chat_history import ChatHistory
+from .complex_prompt import ComplexPrompt
+from .conversation_history import ConversationHistory
+from .intent_classifier import IntentClassifierPrompt
+from .memory import MemoryPlugin
+from .memory_retrieval import MemoryRetrievalPrompt
+from .pii_scrubber import PIIScrubberPrompt
+from .react_prompt import ReActPrompt
+
+__all__ = [
+    "IntentClassifierPrompt",
+    "MemoryRetrievalPrompt",
+    "ReActPrompt",
+    "ComplexPrompt",
+    "ChainOfThoughtPrompt",
+    "ConversationHistory",
+    "MemoryPlugin",
+    "ChatHistory",
+    "PIIScrubberPrompt",
+]
