+++ src/plugins/prompts/chat_history.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from pipeline.plugins.prompts.memory import MemoryPlugin
+
+
+class ChatHistory(MemoryPlugin):
+    """Alias for ``MemoryPlugin`` for backward compatibility."""
+
+    pass
