+++ src/plugins/resources/__init__.py
@@ -1,0 +1,5 @@
+"""Resource plugin re-exports."""
+
+from .cache import CacheResource
+
+__all__ = ["CacheResource"]
