+++ src/plugins/tools/__init__.py
@@ -1,0 +1,9 @@
+from .calculator_tool import CalculatorTool
+from .search_tool import SearchTool
+from .weather_api_tool import WeatherApiTool
+
+__all__ = [
+    "CalculatorTool",
+    "SearchTool",
+    "WeatherApiTool",
+]
